Refactor file-loader into misc
Fix #272
