+++ src/lib/rules/axe/axe.ts
@@ -9,7 +9,7 @@
 
 import * as path from 'path';
 
-import { readFile } from '../../utils/misc';
+import { readFileAsync } from '../../utils/misc';
 import { debug as d } from '../../utils/debug';
 import { ITraverseEnd, IRule, IRuleBuilder, Severity } from '../../types'; // eslint-disable-line no-unused-vars
 import { RuleContext } from '../../rule-context'; // eslint-disable-line no-unused-vars
@@ -54,7 +54,7 @@ const rule: IRuleBuilder = {
 
         const validate = async (traverseEnd: ITraverseEnd) => {
             const { resource } = traverseEnd;
-            const axeCore = await readFile(path.join(process.cwd(), 'node_modules', 'axe-core', 'axe.js'));
+            const axeCore = await readFileAsync(path.join(process.cwd(), 'node_modules', 'axe-core', 'axe.js'));
             const script = `(function () {
     ${axeCore};
     return (${generateScript()}());
